feat(web): 👻 Bag Button
- The Bag Button was refactored to activate the regular modal wallet connect flow.
- The Connect message was refactored to use this component.
